TIP-1159: Move into the right folders
