+++ src/Akeneo/Pim/Enrichment/Bundle/DependencyInjection/AkeneoPimEnrichmentExtension.php
@@ -38,6 +38,7 @@ class AkeneoPimEnrichmentExtension extends Extension
         $loader->load('comparators.yml');
         $loader->load('completeness.yml');
         $loader->load('completeness_checkers.yml');
+        $loader->load('completeness_queries.yml');
         $loader->load('console.yml');
         $loader->load('context.yml');
         $loader->load('converters.yml');
