+++ src/Akeneo/Pim/Enrichment/Bundle/Product/Query/Sql/Completeness/SqlGetProducts.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Component\Product\Completeness;
+namespace Akeneo\Pim\Enrichment\Bundle\Product\Query\Sql\Completeness;
 
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Query\GetProducts;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\Product;
 use Doctrine\DBAL\Connection;
 
 /**
@@ -11,7 +13,7 @@ use Doctrine\DBAL\Connection;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class SqlGetProducts
+final class SqlGetProducts implements \Akeneo\Pim\Enrichment\Component\Product\Completeness\Query\GetProducts
 {
     /** @var Connection */
     private $connection;
