+++ src/Akeneo/Pim/Enrichment/Component/Product/Completeness/Query/GetProducts.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Akeneo\Pim\Enrichment\Component\Product\Completeness\Query;
+
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\Product;
+
+/**
+ * @author    Pierre Allard <pierre.allard@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+interface GetProducts
+{
+    /**
+     * @param string[] $productIdentifiers
+     *
+     * @return Product[]
+     */
+    public function fromProductIdentifiers(array $productIdentifiers): array;
+}
